chore: move files location
