+++ delivery-service/src/main/java/com/sparta/lucky/deliveryservice/domain/delivery/code/DeliveryRouteStatus.java
@@ -1,4 +1,4 @@
-package com.sparta.lucky.deliveryservice.delivery.code;
+package com.sparta.lucky.deliveryservice.domain.delivery.code;
 
 public enum DeliveryRouteStatus {
     WAITING,
