+++ delivery-service/src/main/java/com/sparta/lucky/deliveryservice/domain/driver/code/DriverStatus.java
@@ -1,0 +1,6 @@
+package com.sparta.lucky.deliveryservice.domain.driver.code;
+
+public enum DriverStatus {
+    IDLE,
+    WORKING,
+}
