+++ delivery-service/src/main/java/com/sparta/lucky/deliveryservice/domain/driver/code/DriverType.java
@@ -1,0 +1,6 @@
+package com.sparta.lucky.deliveryservice.domain.driver.code;
+
+public enum DriverType {
+    HUB,
+    COMPANY
+}
