--- delivery-service/src/main/java/com/sparta/lucky/deliveryservice/driver/code/DriverStatus.java
@@ -1,6 +1,0 @@
-package com.sparta.lucky.deliveryservice.driver.code;
-
-public enum DriverStatus {
-    IDLE,
-    WORKING,
-}
