--- delivery-service/src/main/java/com/sparta/lucky/deliveryservice/driver/code/DriverType.java
@@ -1,6 +1,0 @@
-package com.sparta.lucky.deliveryservice.driver.code;
-
-public enum DriverType {
-    HUB,
-    COMPANY
-}
